+++ docs/changes.rst
@@ -19,6 +19,9 @@ Release history
 - fix: attribute :attr:`wxc_sdk.telephony.location.TelephonyLocation.enforce_outside_dial_digit`
 - feat: new :meth:`wxc_sdk.telephony.devices.TelephonyDevicesApi.create_line_key_template`, :meth:`wxc_sdk.telephony.devices.TelephonyDevicesApi.list_line_key_templates`, :meth:`wxc_sdk.telephony.devices.TelephonyDevicesApi.line_key_template_details`, :meth:`wxc_sdk.telephony.devices.TelephonyDevicesApi.modify_line_key_template`, :meth:`wxc_sdk.telephony.devices.TelephonyDevicesApi.delete_line_key_template`, :meth:`wxc_sdk.telephony.devices.TelephonyDevicesApi.preview_apply_line_key_template`
 - feat: improved :meth:wxc_sdk.devices.DevicesApi.list`, use enum parameters
+- fix: :attr:`wxc_sdk.telephony.callqueue.CallQueue.department`, optional
+- fix: :class:`wxc_sdk.common.OwnerType` needs to support PAGING_GROUP and GROUP_PAGING (inconsistent)
+- fix: undocumented attribute :attr:`wxc_sdk.telephone.voicemail_groups.VoicemailGroupDetail.time_zone`
 
 1.16.1
 ------